Create direct test_matmul.py
Use noReductions(logicalDomain) to get number of dimensions in slice
